Fix PR Checks: format, Linux tests, and lint includes
- Format fn_math sources; fix clang-tidy findings (bool masks,
  member init, switch fallthrough, theme color std::array)
- Mark Apple objc_library targets incompatible on non-Apple so
  bazel test //core/... analyzes on Linux CI
- Have tools/lint.sh fetch @pugixml/@luau when headers missing

+++ core/cells/serializer.cc
@@ -190,7 +190,7 @@ void Serializer::serializeTheme(const Workbook& workbook, std::ostream& out) con
     out << "T \"" << escapeString(theme->name) << "\" colors:";
 
     // 12 colors in index order (0-11), no # prefix, comma-separated
-    for (int i = 0; i < 12; ++i) {
+    for (size_t i = 0; i < 12; ++i) {
         if (i > 0) {
             out << ",";
         }
